+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHWC0Y3ETYRTW8MAX1RET7KF"
+        "01KHX02EDWA4JP4JGAA1D78JEQ"
     )
     r"""The unique identifier of the data source"""