+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHWC0Y3HTKGRSEJG6B3TAHP5"
+        "01KHX02EDZZJH9FHPRF52MCENG"
     )
     r"""The unique identifier of the data source"""