+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHWC0Y3PZDSDF56TKY6M88D9"
+        "01KHX02EE3EEDE5A9T7S86S2NT"
     )
     r"""The unique identifier of the data source"""